+++ include/classes/coin_address.class.php
@@ -12,7 +12,7 @@ class CoinAddress extends Base {
    **/
   public function __construct($config) {
     $this->setConfig($config);
-    $this->table = $this->config['db']['shared']['name'] . '.' . $this->table;
+    $this->table = $this->config['db']['shared']['accounts'] . '.' . $this->table;
   }
 
   /**