+++ include/classes/news.class.php
@@ -5,6 +5,17 @@ class News extends Base {
   protected $table = 'news';
 
   /**
+   * We allow changing the database for shared accounts across pools
+   * Load the config on construct so we can assign the DB name
+   * @param config array MPOS configuration
+   * @return none
+   **/
+  public function __construct($config) {
+    $this->setConfig($config);
+    $this->table = $this->config['db']['shared']['news'] . '.' . $this->table;
+  }
+
+  /**
    * Get activation status of post
    * @param id int News ID
    * @return bool true or false
@@ -96,7 +107,7 @@ class News extends Base {
   }
 }
 
-$news = new News();
+$news = new News($config);
 $news->setDebug($debug);
 $news->setMysql($mysqli);
 $news->setUser($user);